Improve filterStates support
FilterStates as emitted by the 'onFilterChanged' event are now encoded
to handle all known cases from the AGGrid docs.

Note that this format is about to change in the future. AGGrid moves
from supporting one or two combined filters (condition1, condition2) to
n (list of conditions). The AgGrid version in this repository at the
time of writing this commit uses 'condition1, condition2'. The latest
versions seem to offer a migration grace period, where both
'condition1,condition2' and 'list condition' are accepted and emitted by
the API.